requirements.txt
asgi-redis==0.12.0
asgiref==1.0.0
autobahn==0.17.1
beautifulsoup4==4.4.0
channels==0.17.3
constantly==15.1.0
daphne==0.15.0
Django==1.10.3
django-appconf==1.0.2
django-imagekit==3.3
gunicorn==19.6.0
incremental==16.10.1
lxml==3.7.1
msgpack-python==0.4.8
pilkit==1.1.13
Pillow==3.4.2
redis==2.10.5
requests==2.7.0
six==1.10.0
Twisted==16.6.0
txaio==2.6.0
whitenoise==3.2.2
zope.interface==4.3.3
